+++ akshare/__init__.py
@@ -2332,9 +2332,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.14 fix: fix option_risk_analysis_em interface
 1.9.15 fix: fix stock_hk_daily interface
 1.9.16 fix: fix stock_financial_abstract interface
+1.9.17 add: add stock_board_industry_spot_em interface
 """
 
-__version__ = "1.9.16"
+__version__ = "1.9.17"
 __author__ = "AKFamily"
 
 import sys
@@ -2740,7 +2741,7 @@ from akshare.stock.stock_board_industry_em import (
     stock_board_industry_hist_em,
     stock_board_industry_hist_min_em,
     stock_board_industry_name_em,
-    stock_board_industry_current_em,
+    stock_board_industry_spot_em,
 )
 
 """