+++ docs/changelog.md
@@ -50,6 +50,10 @@
 
 ## 更新说明
 
+1.9.17 add: add stock_board_industry_spot_em interface
+
+    1. 新增 stock_board_industry_spot_em 接口
+
 1.9.16 fix: fix stock_financial_abstract interface
 
     1. 修复 stock_financial_abstract 接口
@@ -2017,6 +2021,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.17 add: add stock_board_industry_spot_em interface
 
 1.9.16 fix: fix stock_financial_abstract interface